requirements.txt
Flask==3.0.3
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-intel==2.13.0
keras==2.10.0
Keras-Preprocessing==1.1.2
h5py==3.11.0
numpy==1.24.3
matplotlib==3.7.5
Pillow==10.4.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
